packages/server/pyproject.toml
[project]
name = "langgate-server"
version = "0.1.0"
description = "Server implementation for LangGate AI Gateway"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Tanantor Ltd." }]
maintainers = [{ name = "Saran Connoly", email = "saran@tanantor.com" }]
license = { text = "MIT" }
dependencies = [
    "langgate-core",
    "langgate-registry",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pyyaml (>=6.0.1,<7.0.0)",
]

[tool.uv.sources]
langgate-core = { workspace = true }
langgate-registry = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langgate"]
